added more components and hooked refresh cycle back up with event system moved torch inclusion to ml subdirectory

+++ src/include/cxurity/Entity.h
@@ -19,6 +19,6 @@ struct Entity {
   void update();
 
  private:
-  uint8_t upCounter = 0;
+  uint64_t upCounter = 0;
 };
 #endif  //CXURITY_SRC_HEADERS_ENTITY_H_
